requirements.txt
pip install 
django 
djangorestframework 
psycopg2-binary
django-cors-headers
djangorestframework-simplejwt
pillow
openai
dotenv
requests
django-apscheduler
django-scheduler
pytest
pytest-django